requirements.txt
flask==2.3.3
replicate==0.22.0
requests==2.31.0
Pillow==10.0.1
python-dotenv==1.0.0 